pyproject.toml
[tool.poetry]
name = "majocol"
version = "0.1.5"
description = "pick major colors from image"

license = "MIT"

authors = ["suzukey <suzukey28@gmail.com>"]

readme = 'README.md'

repository = "https://github.com/suzukey/majocol"
homepage = "https://github.com/suzukey/majocol"

[tool.poetry.dependencies]
python = "^3.6"
numpy = "^1.19.1"
scikit-learn = "^0.23.2"
Pillow = "^7.2.0"
fitrate = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
isort = "^5.5.1"
mypy = "^0.782"
pytest-cov = "^2.10.1"
black = {version = "^20.8b1", allow-prereleases = true}
flake8 = "^3.8.3"

[tool.black]
line-length = 80
target-version = ['py36', 'py37', 'py38']

[tool.isort]
include_trailing_comma = true
line_length = 80
multi_line_output = 3

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
